requirements.txt
Flask==2.3.2
torch==2.0.1
transformers==4.30.2
scikit-learn==1.3.0
joblib==1.2.0
reportlab==3.6.12
numpy==1.24.3
python-dotenv==1.0.0
sendgrid==6.9.7